Larvel Published

+++ packages/Webkul/Customer/src/Http/Controllers/WishlistController.php
@@ -17,6 +17,13 @@ class WishlistController extends Controller
     protected $_config;
 
     /**
+     * Current customer.
+     *
+     * @var \Webkul\Customer\Models\Customer
+     */
+    protected $currentCustomer;
+
+    /**
      * Create a new controller instance.
      *
      * @param  \Webkul\Customer\Repositories\WishlistRepository  $wishlistRepository
@@ -29,6 +36,8 @@ class WishlistController extends Controller
     )
     {
         $this->_config = request('_config');
+
+        $this->currentCustomer = auth()->guard('customer')->user();
     }
 
     /**
@@ -38,63 +47,58 @@ class WishlistController extends Controller
      */
     public function index()
     {
-        $customer = auth()->guard('customer')->user();
-
         if (! core()->getConfigData('general.content.shop.wishlist_option')) {
             abort(404);
         }
 
         return view($this->_config['view'], [
-            'items'              => $this->wishlistRepository->getCustomerWishlist(),
-            'isSharingEnabled'   => $this->isSharingEnabled(),
-            'isWishlistShared'   => $customer->isWishlistShared(),
-            'wishlistSharedLink' => $customer->getWishlistSharedLink()
+            'items' => $this->wishlistRepository->getCustomerWishlist(),
+            'isSharingEnabled' => $this->isSharingEnabled(),
+            'isWishlistShared' => $this->currentCustomer->isWishlistShared(),
+            'wishlistSharedLink' => $this->currentCustomer->getWishlistSharedLink()
         ]);
     }
 
     /**
      * Function to add item to the wishlist.
      *
-     * @param  int  $productId
+     * @param  int  $itemId
      * @return \Illuminate\Http\Response
      */
-    public function add($productId)
+    public function add($itemId)
     {
-        $customer = auth()->guard('customer')->user();
+        $product = $this->productRepository->findOneByField('id', $itemId);
 
-        $product = $this->productRepository->find($productId);
-
-        if (! $product) {
-            session()->flash('error', trans('customer::app.product-removed'));
-
+        if (! $product->status)
             return redirect()->back();
-        } elseif (! $product->status) {
-            return redirect()->back();
-        }
 
         $data = [
             'channel_id'  => core()->getCurrentChannel()->id,
-            'product_id'  => $productId,
-            'customer_id' => $customer->id,
+            'product_id'  => $itemId,
+            'customer_id' => $this->currentCustomer->id,
         ];
 
-        $wishlist = $this->wishlistRepository->findOneWhere($data);
+        $checked = $this->wishlistRepository->findWhere([
+            'channel_id'  => core()->getCurrentChannel()->id,
+            'product_id'  => $itemId,
+            'customer_id' => $this->currentCustomer->id,
+        ]);
 
-        if (
-            $product->parent
-            && $product->parent->type !== 'configurable'
-        ) {
-            $product = $this->productRepository->find($product->parent_id);
-
+        if ($product->parent && $product->parent->type !== 'configurable') {
+            $product = $this->productRepository->findOneByField('id', $product->parent_id);
             $data['product_id'] = $product->id;
         }
 
-        if (! $wishlist) {
-            $wishlist = $this->wishlistRepository->create($data);
+        if ($checked->isEmpty()) {
+            if ($this->wishlistRepository->create($data)) {
+                session()->flash('success', trans('customer::app.wishlist.success'));
 
-            session()->flash('success', trans('customer::app.wishlist.success'));
+                return redirect()->back();
+            } else {
+                session()->flash('error', trans('customer::app.wishlist.failure'));
 
-            return redirect()->back();
+                return redirect()->back();
+            }
         } else {
             $this->wishlistRepository->findOneWhere([
                 'product_id' => $data['product_id']
@@ -113,22 +117,17 @@ class WishlistController extends Controller
      */
     public function share()
     {
-        $customer = auth()->guard('customer')->user();
-
         if ($this->isSharingEnabled()) {
             $data = $this->validate(request(), [
                 'shared' => 'required|boolean'
             ]);
 
-            $updateCounts = $customer->wishlist_items()->update(['shared' => $data['shared']]);
+            $updateCounts = $this->currentCustomer->wishlist_items()->update(['shared' => $data['shared']]);
 
-            if (
-                $updateCounts
-                && $updateCounts > 0
-            ) {
+            if ($updateCounts && $updateCounts > 0) {
                 return response()->json([
-                    'isWishlistShared'   => $customer->isWishlistShared(),
-                    'wishlistSharedLink' => $customer->getWishlistSharedLink()
+                    'isWishlistShared' => $this->currentCustomer->isWishlistShared(),
+                    'wishlistSharedLink' => $this->currentCustomer->getWishlistSharedLink()
                 ]);
             }
         }
@@ -155,10 +154,7 @@ class WishlistController extends Controller
 
         $items = $customer->wishlist_items()->where('shared', 1)->get();
 
-        if (
-            $customer
-            && $items->isNotEmpty()
-        ) {
+        if ($customer && $items->isNotEmpty()) {
             return view($this->_config['view'], compact('customer', 'items'));
         }
 
@@ -176,10 +172,7 @@ class WishlistController extends Controller
      */
     public function remove($itemId)
     {
-        $customer = auth()->guard('customer')->user();
-
-        $customerWishlistItems = $customer->wishlist_items;
-        $referer = strtok(request()->headers->get('referer'), '?');
+        $customerWishlistItems = $this->currentCustomer->wishlist_items;
 
         foreach ($customerWishlistItems as $customerWishlistItem) {
             if ($itemId == $customerWishlistItem->id) {
@@ -187,7 +180,7 @@ class WishlistController extends Controller
 
                 session()->flash('success', trans('customer::app.wishlist.removed'));
 
-                return redirect()->to($referer);
+                return redirect()->back();
             }
         }
 
@@ -204,11 +197,9 @@ class WishlistController extends Controller
      */
     public function move($itemId)
     {
-        $customer = auth()->guard('customer')->user();
-        
         $wishlistItem = $this->wishlistRepository->findOneWhere([
             'id'          => $itemId,
-            'customer_id' => $customer->id,
+            'customer_id' => $this->currentCustomer->id,
         ]);
 
         if (! $wishlistItem) {
@@ -243,10 +234,12 @@ class WishlistController extends Controller
      */
     public function removeAll()
     {
-        $customer = auth()->guard('customer')->user();
+        $wishlistItems = $this->currentCustomer->wishlist_items;
 
-        foreach ($customer->wishlist_items as $wishlistItem) {
-            $this->wishlistRepository->delete($wishlistItem->id);
+        if ($wishlistItems->count() > 0) {
+            foreach ($wishlistItems as $wishlistItem) {
+                $this->wishlistRepository->delete($wishlistItem->id);
+            }
         }
 
         session()->flash('success', trans('customer::app.wishlist.remove-all-success'));
@@ -261,6 +254,8 @@ class WishlistController extends Controller
      */
     public function isSharingEnabled(): bool
     {
-        return (bool) core()->getConfigData('customer.settings.wishlist.share');
+        return (bool) core()->getConfigData('customer.settings.wishlist.share')
+            ? true
+            : false;
     }
 }
